requirements1.txt
antlr4-python3-runtime==4.9.3
astroid==2.11.7
blessings==1.7
blis==0.7.9
catalogue==1.0.2
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
cymem==2.0.7
Cython==0.29.33
dataclasses==0.6
dill==0.3.5.1
emoji==2.0.0
errant==2.3.3
filelock==3.7.1
gpustat==0.6.0
huggingface-hub==0.8.1
hydra-core==1.3.2
idna==3.3
importlib-resources==5.12.0
isort==5.10.1
joblib==1.1.0
lazy-object-proxy==1.7.1
lxml==4.9.2
mccabe==0.7.0
mock==5.0.1
multiprocess==0.70.13
murmurhash==1.0.9
nltk==3.7
numpy==1.23.1
nvidia-ml-py3==7.352.0
omegaconf==2.2.2
packaging==21.3
pathos==0.2.9
Pillow==9.4.0
plac==1.1.3
platformdirs==2.5.2
portalocker==2.0.0
pox==0.3.1
ppft==1.7.6.5
preshed==3.0.8
protobuf==4.21.2
psutil==5.9.1
pycparser==2.21
pylint==2.14.5
pyparsing==3.0.9
PySnooper==1.1.1
PyYAML==6.0
rapidfuzz==2.13.7
regex==2022.7.9
requests==2.28.1
sacrebleu==1.5.1
six==1.16.0
spacy==2.3.9
srsly==1.0.6
stanza==1.4.0
subword-nmt==0.3.8
supar==1.1.4
tabulate==0.9.0
thinc==7.4.6
tokenizers==0.12.1
TorchSnooper==0.8
tqdm==4.65.0
transformers==4.20.1
typing_extensions==4.3.0
urllib3==1.26.10
wasabi==0.10.1
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
